Adjust styling on calculator
Readout text color was black, making it very difficult to read on the
black screen. This commit fixes this.
Additionally, I adjusted where on
the screen the calculator sits, centering it rather than the default
positioning of the far left.
I also made a slight change in the background color of the overall site
to an off-white color.