Adjust values on operator buttons
While considering how to display the formula being calculated, I
realized the best process would be to just display the value of each
button. This was a problem for the operators (add/subtract/multiply/
divide) as we would want them displayed by their symbols (+/-/x/'/').
This commit changes the values on each of the operator buttons to match
the symbol rather than the operation name.