Add display function
This commit adds a function to display the value of each button to the
readout of the calculator.
Additionally, I changed how each button is called. Rather than having a
const for each number, I have a single buttons const that when pushed,
calls on the value of the buttons set on the html file. This will make
all values easier to call in to javascript.