Adjust visuals of readout screen & buttons.
I realized the readout screen was a bit difficult to read. To resolve,
I adjusted readout screen and buttons to use a larger and easier to read
font. I also set the readout screen to add text to the vertical center
of the div, rather than at the top. I also changed it to align the
readout display on the right rather than the left. This last change was
purely for aesthetics.