Add event listeners to set operator
When any operator is clicked, there is now an event listener to set the
selected operator to be the operator variable value in javascript.
I also in this commit moved the way operators are displayed in the
readout. They were previously logged via the value in the same way that
numbers are logged. The problem with this was that the operators were
also being logged to the firstNumber array, causing issues when it is
then turned later into an integer. For example, we had arrays along the
lines of '[5,6,+]' rather than '[5,6]'.