Remove values from operators
In the last commit, I moved the way operators are displayed in the
readout. They were previously logged via the value in the same way that
numbers are logged. The problem with this was that the operators were
also being logged to the firstNumber array, causing issues when it is
then turned later into an integer. For example, we had arrays along the
lines of '[5,6,+]' rather than '[5,6]'.
To complete this resolution, this commit removes those values from the
operator buttons on the html file.