Change equation Display from string to array
This commit changes the equation display to an array rather than a
string. I chose an array over a string because with an array, I will
have the option to pop out previous values. For example, if someone
clicks + (plus) when they meant - (minus), the string version of the
equation readout would reflect '+-' (plus and minus side by side),
rather than the ideal outcome of the plus sign being removed.
I also in this commit set a default display and first number of 0. This
will prevent NaN answers when firstNumber is left blank, i.e. '+6'
(plus six).