Create function to Operate and Reset
This commit adds the operateAndReset function. This function combines
the existing operate function with the necessary steps to also reset the
calculator for the next equation. More specifically, after the operate
function, it sets the returned value as the firstNumber variable,
it also clears the secondNumber and operator variable. Lastly, it clears
the readout to just read the returned value.
I made this into it's own function rather than simply calling all needed
actions when equals is clicked, because I also want to use this function
when a second operator is clicked after there is a secondNumber present.
For example, if a user inputs '6+2-4' (six plus two minus four). Since
this calculator is more simple, such equations will be handled in parts.
Using the same example, '6+2=8' (six plus two equals eight),
followed by '8-4' (eight minus four).