Correct readout and decimal equations
The aim in these changes was to correct the readout screen when an
operator was clicked then changed before secondNumber was entered.
To resolve this, I changed the displayValue function. Rather than simply
pushing any entry to the end of the display, the display reads out the
firstNumber value then the operator value then the secondNumber value.
This fixed multiple operators appearing side by side, now if the
operator is changed, the old operator disappears and is replaced by the
new operator.
In this resolution, I noticed that if firstNumber or secondNumber were
2 digits or more, they would appear with commas between them as they are
arrays. I could fix this in the same way I had been:
'parseInt(firstNumber.join('')). However, I realized this was being used
numerous times and would be better as it's own formula, so I made the
correctNumbers() formula.
I had also noticed in testing that all decimal values would be rounded
down. For example '1.5 x 2 = 2' (one-point-five times two equals two).
I found this was because I should be using parseFloat rather than
parseInt. While building the new formula, I went ahead and made this
change.
Lastly, I updated all prior uses of parseInt() to instead use the new
formula.